split qwen36 moe bake bootstrap
