+++ crates/runner/src/main.rs
@@ -13,6 +13,7 @@ mod policy;
 mod prefill_engine;
 mod qwen35_dflash_engine;
 mod qwen35_runtime;
+mod qwen36_moe_bake;
 mod qwen36_moe_decode;
 mod qwen36_moe_dry_run;
 mod qwen36_moe_engine;
